OCT-111: moved GetFamiliesByCodeQuery
