+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/services.yml
@@ -72,7 +72,7 @@ services:
         arguments:
             - '@pim_enrich.repository.family.search'
 
-    Akeneo\Catalogs\Infrastructure\Persistence\GetFamiliesByCodeQuery:
+    Akeneo\Catalogs\Infrastructure\Persistence\Family\GetFamiliesByCodeQuery:
         arguments:
             - '@pim_enrich.repository.family.search'
